# Conflicts:
#	nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.dom/src/main/java/nts/uk/ctx/at/request/dom/application/applist/service/content/AppContentServiceImpl.java
#	nts.uk/uk.at/at.web/nts.uk.at.web/src/main/webapp/view/kaf/002/m/kaf002.m.vm.ts
#	nts.uk/uk.at/at.web/nts.uk.at.web/src/main/webapp/view/kaf/009/a/kaf009.a.vm.ts
#	nts.uk/uk.at/at.web/nts.uk.at.web/src/main/webapp/view/kaf/009/shr/kaf009.shr.vm.ts
#	nts.uk/uk.mobile/ClientApp/views/kaf/s07/a/index.ts

+++ nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.dom/src/main/java/nts/uk/ctx/at/request/dom/application/ApplicationApprovalImpl.java
@@ -12,6 +12,7 @@ import nts.uk.ctx.at.request.dom.application.businesstrip.BusinessTrip;
 import nts.uk.ctx.at.request.dom.application.businesstrip.BusinessTripRepository;
 import nts.uk.ctx.at.request.dom.application.common.adapter.workflow.ApprovalRootStateAdapter;
 import nts.uk.ctx.at.request.dom.application.common.adapter.workflow.dto.ApprovalPhaseStateImport_New;
+import nts.uk.ctx.at.request.dom.application.gobackdirectly.GoBackDirectlyRepository;
 import nts.uk.ctx.at.request.dom.application.gobackdirectly.GoBackDirectlyRepository_Old;
 import nts.uk.ctx.at.request.dom.application.holidayshipment.absenceleaveapp.AbsenceLeaveAppRepository;
 import nts.uk.ctx.at.request.dom.application.holidayshipment.brkoffsupchangemng.BrkOffSupChangeMng;
@@ -20,7 +21,11 @@ import nts.uk.ctx.at.request.dom.application.holidayshipment.recruitmentapp.Recr
 import nts.uk.ctx.at.request.dom.application.holidayworktime.AppHolidayWorkRepository;
 import nts.uk.ctx.at.request.dom.application.lateleaveearly.ArrivedLateLeaveEarlyRepository;
 import nts.uk.ctx.at.request.dom.application.overtime.OvertimeRepository;
+import nts.uk.ctx.at.request.dom.application.stamp.AppRecordImage;
+import nts.uk.ctx.at.request.dom.application.stamp.AppRecordImageRepository;
+import nts.uk.ctx.at.request.dom.application.stamp.AppStampRepository;
 import nts.uk.ctx.at.request.dom.application.stamp.AppStampRepository_Old;
+import nts.uk.ctx.at.request.dom.application.stamp.StampRequestMode;
 import nts.uk.ctx.at.request.dom.application.workchange.AppWorkChangeRepository;
 import nts.uk.shr.com.context.AppContexts;
 
@@ -40,13 +45,16 @@ public class ApplicationApprovalImpl implements ApplicationApprovalService {
 	private ApprovalRootStateAdapter approvalRootStateAdapter;
 
 	@Inject
-	private AppStampRepository_Old appStampRepository;
+	private AppStampRepository appStampRepository;
+	
+	@Inject
+	private AppRecordImageRepository appRecordImageRepository;
 
 	@Inject
 	private OvertimeRepository overtimeRepository;
 
 	@Inject
-	private GoBackDirectlyRepository_Old goBackDirectlyRepository;
+	private GoBackDirectlyRepository goBackDirectlyRepository;
 
 	@Inject
 	private AppWorkChangeRepository workChangeRepository;
@@ -77,7 +85,13 @@ public class ApplicationApprovalImpl implements ApplicationApprovalService {
 		Application application = applicationRepository.findByID(appID).get();
 		switch (application.getAppType()) {
 		case STAMP_APPLICATION:
-			appStampRepository.delete(companyID, appID);
+			if (application.getOpStampRequestMode().isPresent()) {
+				if (application.getOpStampRequestMode().get() == StampRequestMode.STAMP_ADDITIONAL) {
+					appStampRepository.delete(companyID, appID);					
+				} else {
+					appRecordImageRepository.delete(companyID, appID);
+				}
+			}
 			break;
 		case OVER_TIME_APPLICATION:
 			overtimeRepository.delete(companyID, appID);
